burger/requirements.txt
django 
djangorestframework
django-cors-headers
djangorestframework-expiring-authtoken